requirements.txt
Flask==2.1.2
Flask_Cors==3.0.10
websockets==10.3
pymongo==4.2.0
python-dotenv==0.20.0